fix(app): a CDN entry is TLS, not REALITY, and the builders assumed otherwise
The partner subscription's CDN row is xhttp over ordinary TLS to a host with
a real certificate — `security=tls`, `mode=stream-one`, no `pbk`. Both config
builders emitted a REALITY block regardless, so Xray was handed an empty key
and said so exactly:

    Failed to build REALITY config > empty "password"

which reads as a missing credential rather than as the wrong kind of security
altogether. The other six xhttp rows carry a key and were fine, which is why
this only ever showed up on one row out of seven.

`XrayConfig.buildXhttp` now emits `tlsSettings` when there is no public key
and `realitySettings` when there is. `SingBoxConfig` gets the same rule: a
reality block with an empty key is not a weaker handshake, it is a rejected
config. `TransportKind.Tls` already existed to name this case; nothing was
reading it. Three tests, one per branch.

Warnings, while here. `-Xexpect-actual-classes` accepts what the compiler
asks for on every build (KT-61573) — `PlatformCrypto` is an `expect object`
on purpose. A redundant `.toInt()` in the Apple crypto is gone.

`rememberModalBottomSheetState` is suppressed at its four sites rather than
migrated: its replacement is itself alpha in the pinned material3, and
swapping a sheet API blind, mid-release, on a build nobody here can compile is
a worse trade than a warning. Worth doing deliberately, all four together.

+++ sharedUI/src/appleMain/kotlin/org/olcbox/app/crypt/PlatformCrypto.apple.kt
@@ -81,7 +81,7 @@ actual object PlatformCrypto {
                             cp.addressOf(0), ciphertext.size.convert(),
                             op.addressOf(0), outBuf.size.convert(),
                             moved.ptr
-                        ).toInt()
+                        )
                     }
                 }
             }
